arvolaskuri/requirements.txt
# pip install -r requirements.txt
# pip install --upgrade --user google-cloud-aiplatform
# pip install google-cloud-vision  
# gcloud auth application-default login

argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
Django==5.0.2
djangorestframework==3.14.0
drf-extra-fields==3.7.0
filetype==1.2.0
google-api-core==2.17.1
google-auth==2.28.1
google-cloud-aiplatform==1.43.0
google-cloud-bigquery==3.17.2
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.2
google-cloud-storage==2.14.0
google-cloud-vision==3.7.1
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.0
grpcio-status==1.62.0
idna==3.6
numpy==1.26.4
packaging==23.2
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
rsa==4.9
shapely==2.0.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
python-keystoneclient==5.4.0
python-swiftclient==4.5.0